[#13] Revert to uint256 type for amount param

+++ contracts/Bridge.sol
@@ -424,7 +424,7 @@ contract Bridge {
     string calldata source,
     string calldata issuer,
     uint64 destinationTag,
-    uint64 amount
+    uint256 amount
   ) internal {
     bytes32 currencyHash = keccak256(abi.encodePacked(currencyCode, issuer));
